+++ packages/ui/src/components/Chip/ChipItem.tsx
@@ -1,0 +1,54 @@
+import { ComponentPropsWithoutRef, forwardRef, ReactElement } from 'react';
+import { chipCloseButtonRecipe, chipRecipe } from './Chip.css';
+import { Icon } from '../Icon/Icon';
+import { Text } from '../Text/Text';
+import { isNil } from '../../utils';
+
+export type ButtonVariant = 'grey' | 'purple' | 'green';
+
+export type ChipProps = ComponentPropsWithoutRef<'span'> & {
+  variant: ButtonVariant;
+  leftAddon?: ReactElement;
+  rightAddon?: ReactElement;
+  closable?: boolean;
+  onClose?: () => void;
+};
+
+export const ChipItem = forwardRef<HTMLSpanElement, ChipProps>(
+  ({
+    className = '',
+    variant,
+    leftAddon,
+    rightAddon,
+    closable = false,
+    children,
+    onClose,
+    ...rest
+  }) => {
+    return (
+      <span className={`${chipRecipe({ variant })} ${className}`} {...rest}>
+        {leftAddon}
+        <Text fontSize={16} fontWeight="semibold">
+          {children}
+        </Text>
+        {rightAddon}
+        {isNil(closable) && (
+          <button
+            type="button"
+            aria-label="Close"
+            className={chipCloseButtonRecipe({ clickable: closable })}
+            onClick={onClose}
+          >
+            <Icon
+              name="x"
+              size="1.6rem"
+              type="stroke"
+              strokeWidth={'0.244rem'}
+              color={`${variant}400`}
+            />
+          </button>
+        )}
+      </span>
+    );
+  }
+);
